requirements.txt
openai>=1.40
python-dotenv>=1.0.1
PyYAML>=6.0.2
tqdm>=4.66.5
